expand range of item disambiguated name level

+++ src/ir/item.rs
@@ -711,7 +711,7 @@ impl Item {
         &self,
         ctx: &BindgenContext,
         to: &mut String,
-        level: u8,
+        level: u16,
     ) {
         to.push_str(&self.canonical_name(ctx));
         if let ItemKind::Type(ref ty) = *self.kind() {
